backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
tensorflow==2.15.0
pymongo==4.6.1
python-dotenv==1.0.1
gunicorn==21.2.0
structlog==23.2.0
reportlab==4.0.7
bcrypt==4.1.2
PyJWT==2.8.0
email-validator==2.2.0
